requirements.txt
ccxt>=4.2.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
pyyaml>=6.0
pyarrow>=12.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
